+++ com/pdfjet/OpenTypeFont.java
@@ -33,6 +33,7 @@ class OpenTypeFont {
         font.markToMarkOffsets = otf.markToMarkOffsets;
         font.markAnchors = otf.markAnchors;
         font.baseAnchors = otf.baseAnchors;
+        font.checksum = Font.checksumOf(font);
         font.setSize(font.size);
 
         embedFontFile(pdf, font, otf);
@@ -67,7 +68,7 @@ class OpenTypeFont {
     private static void embedFontFile(PDF pdf, Font font, OTF otf) throws Exception {
         // Check if the font file is already embedded
         for (Font f : pdf.fonts) {
-            if (f.fileObjNumber != 0 && f.name.equals(otf.fontName)) {
+            if (f.fileObjNumber != 0 && f.name.equals(otf.fontName) && f.checksum == font.checksum) {
                 font.fileObjNumber = f.fileObjNumber;
                 return;
             }
@@ -114,7 +115,7 @@ class OpenTypeFont {
     private static void addFontDescriptorObject(
             PDF pdf, Font font, OTF otf) throws Exception {
         for (Font f : pdf.fonts) {
-            if (f.fontDescriptorObjNumber != 0 && f.name.equals(otf.fontName)) {
+            if (f.fontDescriptorObjNumber != 0 && f.name.equals(otf.fontName) && f.checksum == font.checksum) {
                 font.fontDescriptorObjNumber = f.fontDescriptorObjNumber;
                 return;
             }
@@ -175,7 +176,7 @@ class OpenTypeFont {
             Font font,
             OTF otf) throws Exception {
         for (Font f : pdf.fonts) {
-            if (f.toUnicodeCMapObjNumber != 0 && f.name.equals(otf.fontName)) {
+            if (f.toUnicodeCMapObjNumber != 0 && f.name.equals(otf.fontName) && f.checksum == font.checksum) {
                 font.toUnicodeCMapObjNumber = f.toUnicodeCMapObjNumber;
                 return;
             }
@@ -256,7 +257,7 @@ class OpenTypeFont {
             Font font,
             OTF otf) throws Exception {
         for (Font f : pdf.fonts) {
-            if (f.cidFontDictObjNumber != 0 && f.name.equals(otf.fontName)) {
+            if (f.cidFontDictObjNumber != 0 && f.name.equals(otf.fontName) && f.checksum == font.checksum) {
                 font.cidFontDictObjNumber = f.cidFontDictObjNumber;
                 return;
             }